questionpapergenerator-main/site/api/requirements.txt
fastapi[all]
uvicorn
python-dotenv
requests
pypdf
python-multipart